requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai
langchain
faiss-cpu
whisper
